Move SchemaExplorer to shared dir
