+++ ui/src/dashboards/components/QueryMaker.js
@@ -3,7 +3,7 @@ import React, {PropTypes} from 'react'
 import EmptyQuery from 'src/shared/components/EmptyQuery'
 import QueryTabList from 'src/shared/components/QueryTabList'
 import QueryTextArea from 'src/dashboards/components/QueryTextArea'
-import SchemaExplorer from 'src/dashboards/components/SchemaExplorer'
+import SchemaExplorer from 'src/shared/components/SchemaExplorer'
 import buildInfluxQLQuery from 'utils/influxql'
 
 const TEMPLATE_RANGE = {upper: null, lower: ':dashboardTime:'}
